Add more example module for readwrite-splitting
