+++ examples/shardingsphere-sample/shardingsphere-jdbc-sample/shardingsphere-jdbc-memory-example/shardingsphere-jdbc-memory-local-example/shardingsphere-jdbc-memory-local-readwrite-splitting-example/shardingsphere-jdbc-memory-local-readwrite-splitting-springboot-starter-mybatis-example/pom.xml
@@ -20,22 +20,22 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>shardingsphere-jdbc-memory-local-example</artifactId>
+        <artifactId>shardingsphere-jdbc-memory-local-readwrite-splitting-example</artifactId>
         <groupId>org.apache.shardingsphere.example</groupId>
         <version>5.0.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
-    <artifactId>shardingsphere-jdbc-memory-local-readwrite-splitting-example</artifactId>
+    <artifactId>shardingsphere-jdbc-memory-local-readwrite-splitting-springboot-starter-mybatis-example</artifactId>
     <name>${project.artifactId}</name>
     
-    <modules>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-jdbc-example</module>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-spring-namespace-jdbc-example</module>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-springboot-starter-jdbc-example</module>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-springboot-starter-jpa-example</module>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-spring-namespace-jpa-example</module>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-spring-namespace-mybatis-example</module>
-        <module>shardingsphere-jdbc-memory-local-readwrite-splitting-springboot-starter-mybatis-example</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>shardingsphere-jdbc-core-spring-boot-starter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mybatis.spring.boot</groupId>
+            <artifactId>mybatis-spring-boot-starter</artifactId>
+        </dependency>
+    </dependencies>
 </project>
