--- examples/shardingsphere-sample/shardingsphere-jdbc-sample/shardingsphere-jdbc-memory-example/shardingsphere-jdbc-memory-local-example/shardingsphere-jdbc-memory-local-sharding-example/shardingsphere-jdbc-memory-local-sharding-springboot-starter-mybatis-example/src/main/java/org/apache/shardingsphere/example/sharding/springboot/starter/mybatis/entity/Order.java
@@ -1,70 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.example.sharding.springboot.starter.mybatis.entity;
-
-import java.io.Serializable;
-
-public class Order implements Serializable {
-    
-    private static final long serialVersionUID = 8306802022239174861L;
-    
-    private long orderId;
-    
-    private int userId;
-    
-    private long addressId;
-    
-    private String status;
-    
-    public long getOrderId() {
-        return orderId;
-    }
-    
-    public void setOrderId(final long orderId) {
-        this.orderId = orderId;
-    }
-    
-    public int getUserId() {
-        return userId;
-    }
-    
-    public void setUserId(final int userId) {
-        this.userId = userId;
-    }
-    
-    public String getStatus() {
-        return status;
-    }
-    
-    public void setStatus(final String status) {
-        this.status = status;
-    }
-    
-    public long getAddressId() {
-        return addressId;
-    }
-    
-    public void setAddressId(final long addressId) {
-        this.addressId = addressId;
-    }
-    
-    @Override
-    public String toString() {
-        return String.format("order_id: %s, user_id: %s, address_id: %s, status: %s", orderId, userId, addressId, status);
-    }
-}
